Move generation parameters to data
