+++ hypothesis-python/src/hypothesis/internal/conjecture/data.py
@@ -17,6 +17,7 @@
 
 from __future__ import absolute_import, division, print_function
 
+import math
 from collections import defaultdict
 from enum import IntEnum
 
@@ -33,7 +34,7 @@ from hypothesis.internal.compat import (
     text_type,
     unicode_safe_repr,
 )
-from hypothesis.internal.conjecture.junkdrawer import IntList
+from hypothesis.internal.conjecture.junkdrawer import IntList, uniform
 from hypothesis.internal.conjecture.utils import calc_label_from_name
 from hypothesis.internal.escalation import mark_for_escalation
 
@@ -1052,3 +1053,135 @@ def bits_to_bytes(n):
     Equivalent to (n + 7) // 8, but slightly faster. This really is
     called enough times that that matters."""
     return (n + 7) >> 3
+
+
+generation_parameters_count = 0
+
+
+class GenerationParameters(object):
+    """Parameters to control generation of examples."""
+
+    AVERAGE_ALPHABET_SIZE = 3
+
+    ALPHABET_FACTOR = math.log(1.0 - 1.0 / AVERAGE_ALPHABET_SIZE)
+
+    def __init__(self, random):
+        self.__random = random
+        self.__zero_chance = None
+        self.__max_chance = None
+        self.__pure_chance = None
+        self.__alphabet = {}
+
+        global generation_parameters_count
+        generation_parameters_count += 1
+
+        self.__id = generation_parameters_count
+
+    def __repr__(self):
+        return "GenerationParameters(%d)" % (self.__id,)
+
+    def draw_bytes(self, n):
+        """Draw an n-byte block from the distribution defined by this instance
+        of generation parameters."""
+        alphabet = self.alphabet(n)
+
+        if alphabet is None:
+            return self.__draw_without_alphabet(n)
+
+        return self.__random.choice(alphabet)
+
+    def __draw_without_alphabet(self, n):
+        if self.__random.random() <= self.zero_chance:
+            return hbytes(n)
+
+        if self.__random.random() <= self.max_chance:
+            return hbytes([255]) * n
+
+        return uniform(self.__random, n)
+
+    def alphabet(self, n_bytes):
+        """Returns an alphabet - a list of values to use for all blocks with
+        this number of bytes - or None if this value should be generated
+        without an alphabet.
+
+        This is designed to promote duplication in the test case that would
+        otherwise happen with very low probability.
+        """
+        try:
+            return self.__alphabet[n_bytes]
+        except KeyError:
+            pass
+
+        if self.__random.random() <= self.pure_chance:
+            # Sometiems we don't want to use an alphabet (e.g. for generating
+            # sets of integers having a small alphabet is disastrous), so with
+            # some probability we want to generate choices that do not use the
+            # alphabet. As with other factors we set this probability globally
+            # across the whole choice of distribution so we have various levels
+            # of mixing.
+            result = None
+        else:
+            # We draw the size as a geometric distribution with average size
+            # GenerationParameters.AVERAGE_ALPHABET_SIZE.
+            size = (
+                int(
+                    math.log(self.__random.random())
+                    / GenerationParameters.ALPHABET_FACTOR
+                )
+                + 1
+            )
+            assert size > 0
+
+            size = self.__random.randint(1, 10)
+            result = [self.__draw_without_alphabet(n_bytes) for _ in hrange(size)]
+
+        self.__alphabet[n_bytes] = result
+        return result
+
+    @property
+    def max_chance(self):
+        """Returns a probability with which any given draw_bytes call should
+        be forced to be all 255 bytes. This is an important value because it
+        can make it more likely to push us into rare corners of the search
+        space, especially when biased_coin is being used."""
+        if self.__max_chance is None:
+            # We want to generate pure random examples every now and then. This
+            # is partly to offset too strong a bias to zero and partly because
+            # some user defined strategies may not play well with zero biasing.
+            self.__max_chance = self.__random.random() * 0.01
+        return self.__max_chance
+
+    @property
+    def zero_chance(self):
+        """Returns a probability with which any given draw_bytes call should
+        be forced to be all zero. This is an important value especially because
+        it will tend to force the test case to be smaller than it otherwise
+        would be."""
+        if self.__zero_chance is None:
+            # We want to generate pure random examples every now and then. This
+            # is partly to offset too strong a bias to zero and partly because
+            # some user defined strategies may not play well with zero biasing.
+            if self.__random.randrange(0, 10) == 0:
+                self.__zero_chance = 0.0
+            else:
+                self.__zero_chance = self.__random.random() * 0.5
+        return self.__zero_chance
+
+    @property
+    def pure_chance(self):
+        """Returns a probability with which any given draw_bytes call should
+        be forced to be all pure."""
+        if self.__pure_chance is None:
+            self.__pure_chance = self.__random.random()
+        return self.__pure_chance
+
+
+def draw_bytes_with(prefix, parameter):
+    def draw_bytes(data, n):
+        if data.index < len(prefix):
+            result = prefix[data.index : data.index + n]
+            # We always draw prefixes as a whole number of blocks
+            return result
+        return parameter.draw_bytes(n)
+
+    return draw_bytes
